requirements.txt
autopep8==1.7.0
requests
pandas
opencv-python==4.6.0.66
panda==0.3.1
pandas==1.5.1
python-dotenv
requests==2.28.1
